IZZO PROYECT/requirements.txt
pytest
pytest-flask

